+++ tests/BusinessLogic/ShippingMethod/ShippingMethodServiceTest.php
@@ -160,16 +160,18 @@ class ShippingMethodServiceTest extends BaseTestWithServices
     public function testActivateMethod()
     {
         $serviceDetails = $this->getShippingServiceDetails(1);
+        $secondDetails = $this->getShippingServiceDetails(2, 'new_carier');
         $method = $this->shippingMethodService->add($serviceDetails);
+        $method2 = $this->shippingMethodService->add($secondDetails);
 
         self::assertCount(0, $this->shippingMethodService->getActiveMethods());
-        self::assertCount(1, $this->shippingMethodService->getAllMethods());
+        self::assertCount(2, $this->shippingMethodService->getAllMethods());
 
         $this->shippingMethodService->activate($method->getId());
 
         self::assertTrue($this->shippingMethodService->isAnyMethodActive());
         self::assertCount(
-            1,
+            2,
             $this->testShopShippingMethodService->callHistory,
             'Activation should be triggered in shop!'
         );
@@ -180,11 +182,31 @@ class ShippingMethodServiceTest extends BaseTestWithServices
             'Activation should be triggered in shop!'
         );
 
+        self::assertCount(
+            1,
+            $this->testShopShippingMethodService->callHistory['addBackup'],
+            'Backup shipping method should be added in shop.'
+        );
+
+        $this->shippingMethodService->activate($method2->getId());
+
+        self::assertCount(
+            2,
+            $this->testShopShippingMethodService->callHistory['add'],
+            'Activation should be triggered in shop!'
+        );
+
+        self::assertCount(
+            1,
+            $this->testShopShippingMethodService->callHistory['addBackup'],
+            'Backup shipping method should not be added in shop if there are already active methods.'
+        );
+
         $serviceDetails->serviceName = 'changed name';
         $this->shippingMethodService->update($serviceDetails);
 
         self::assertCount(
-            2,
+            3,
             $this->testShopShippingMethodService->callHistory,
             'Update should be propagated to shop when instance is active!'
         );
@@ -192,6 +214,11 @@ class ShippingMethodServiceTest extends BaseTestWithServices
             1,
             $this->testShopShippingMethodService->callHistory['update'],
             'Update should be propagated to shop when instance is active!'
+        );
+        self::assertCount(
+            1,
+            $this->testShopShippingMethodService->callHistory['addBackup'],
+            'Backup shipping method should not be added in shop on update.'
         );
     }
 
@@ -211,6 +238,53 @@ class ShippingMethodServiceTest extends BaseTestWithServices
             1,
             $this->testShopShippingMethodService->callHistory['delete'],
             'Deactivation should be triggered in shop!'
+        );
+
+        self::assertCount(
+            1,
+            $this->testShopShippingMethodService->callHistory['deleteBackup'],
+            'Deleting backup shipping should be triggered in shop!'
+        );
+
+        self::assertCount(0, $this->shippingMethodService->getActiveMethods());
+
+        $method2 = $this->shippingMethodService->add($this->getShippingServiceDetails(2, 'second_carrier'));
+
+        self::assertCount(2, $this->shippingMethodService->getAllMethods());
+
+        $this->shippingMethodService->activate($method->getId());
+        $this->shippingMethodService->activate($method2->getId());
+
+        self::assertCount(2, $this->shippingMethodService->getActiveMethods());
+
+        $this->shippingMethodService->deactivate($method->getId());
+
+        self::assertCount(1, $this->shippingMethodService->getActiveMethods());
+
+        self::assertCount(
+            2,
+            $this->testShopShippingMethodService->callHistory['delete'],
+            'Deactivation should be triggered in shop!'
+        );
+
+        self::assertCount(
+            1,
+            $this->testShopShippingMethodService->callHistory['deleteBackup'],
+            'Deleting backup shipping should not be triggered in shop when there are active methods!'
+        );
+
+        $this->shippingMethodService->deactivate($method2->getId());
+
+        self::assertCount(
+            3,
+            $this->testShopShippingMethodService->callHistory['delete'],
+            'Deactivation should be triggered in shop!'
+        );
+
+        self::assertCount(
+            2,
+            $this->testShopShippingMethodService->callHistory['deleteBackup'],
+            'Deleting backup shipping should be triggered in shop!'
         );
 
         self::assertCount(0, $this->shippingMethodService->getActiveMethods());
@@ -316,12 +390,12 @@ class ShippingMethodServiceTest extends BaseTestWithServices
         );
     }
 
-    private function getShippingServiceDetails($id)
+    private function getShippingServiceDetails($id, $carrierName = 'test carrier')
     {
         $details = ShippingServiceDetails::fromArray(
             array(
                 'id' => $id,
-                'carrier_name' => 'test carrier',
+                'carrier_name' => $carrierName,
                 'service_name' => 'test service',
                 'currency' => 'EUR',
                 'country' => 'IT',